rename command to command menu
