+++ app/layout.tsx
@@ -1,5 +1,5 @@
-import { CommandMenu } from '@/components/command';
-import { CommandMenuProvider } from '@/components/command/command-menu-context';
+import { CommandMenu } from '@/components/command-menu';
+import { CommandMenuProvider } from '@/components/command-menu/command-menu-context';
 import InteractiveConfigContextWrapper from '@/components/idiot/interactive/ContextWrapper';
 import { SidebarContentProvider } from '@/components/layout/SidebarContentManager';
 import { SidebarContext } from '@/components/layout/SidebarContext';
@@ -13,7 +13,7 @@ import { cookies } from 'next/headers';
 import { ReactNode } from 'react';
 import './globals.css';
 import { metadata, viewport } from './metadata';
-import { SolanaWalletProvider } from '@/components/wallet/wallet-provider';
+import { SolanaWalletProvider } from '@/components/command-menu/wallet-provider';
 
 const inter = Inter({ subsets: ['latin'] });
