+++ components/command-menu/group/chat-history.tsx
@@ -3,7 +3,7 @@
 import dayjs from 'dayjs';
 import { useCallback } from 'react';
 import { useRouter } from 'next/navigation';
-import { useCommandMenu } from '@/components/command/command-menu-context';
+import { useCommandMenu } from '@/components/command-menu/command-menu-context';
 import { useConversations } from '@/components/idiot/interactive/hooks/useConversation';
 import { CommandGroup, CommandItem } from '@/components/ui/command';
